+++ src/app/(routes)/food/FoodList.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { MenuContext } from "../../context/MenuContext";
 import { TMenuItemFood } from "../../../interfaces/menuItem";
-import { MenuItemFood } from "../../components/MenuItemFood";
+import { MenuItemFood } from "../../components/ListItemFood";
 
 const FoodList: React.FC = () => {
 	const [dishes, setDishes] = useState<TMenuItemFood[]>([]);
@@ -23,8 +23,6 @@ const FoodList: React.FC = () => {
 		<div className="foodList">
 			<h1>Our Dishes</h1>
 			<hr />
-			{dishes.length}hallo
-			{dishes.length > 0 ? dishes.length : "No dishes"}
 			<div className="menuItems">
 				{dishes &&
 					dishes.length > 0 &&
